requirements.txt
atlassian-python-api==3.13.0
lxml==4.6.3
pandas==1.3.1
plotly==5.2.1
kaleido==0.2.1